aide-data-pipeline/pyproject.toml
[tool.poetry]
name = "aide-data-pipeline"
version = "0.1.0"
description = "AIDE Platform Data Pipeline Orchestrator - Unified data collection and processing"
authors = ["AIDE Team <team@aide-platform.com>"]
readme = "README.md"
packages = [{include = "aide_pipeline"}]

[tool.poetry.dependencies]
python = "^3.11"
aide-data-core = {path = "../aide-data-core", develop = true}
aide-crawlers = {path = "../aide-crawlers", develop = true}
aide-data-engine = {path = "../aide-data-engine", develop = true}
pyyaml = "^6.0.0"
typer = {extras = ["all"], version = "^0.9.0"}
rich = "^13.7.0"
schedule = "^1.2.0"
python-dotenv = "^1.0.0"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.0"
pytest-cov = "^4.1.0"
black = "^23.7.0"
isort = "^5.12.0"
mypy = "^1.4.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 100
target-version = ['py311']

[tool.isort]
profile = "black"
line_length = 100

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
